mt6755: Fix lcd backlight triggering
The brightness can only be set when the display is powered up. This
patch turns off the backlight before the display is suspended and turns
it on after it is powered up.

It also fixes the lcd backlight brightness initial value, the backlight
is initialized by the bootloader but the actual brightness value is not
set properly in the led device.

Change-Id: Ia67bf1c1f26385379b1a0f922a27bf24066dc448